empanaderia_chilena/requirements.txt
fastapi
uvicorn
pymongo
pydantic
passlib[bcrypt]
pyjwt
requests
pytest
httpx
python-dotenv
aiofiles